add pytest

+++ tests/conftest.py
@@ -1,17 +1,25 @@
+import os
 from typing import AsyncGenerator
 from httpx import ASGITransport, AsyncClient
 
 import pytest_asyncio
 
+from src.config import settings
 from src.database import engine
 from src.main import app
 from src.models import Base
 
+@pytest_asyncio.fixture(scope="session", autouse=True)
+def set_test_mode():
+    print(f"MODE: {settings.MODE}")
+    settings.MODE = "TEST"
+    yield
+    settings.MODE = "TEST"
 
 @pytest_asyncio.fixture(scope="session", autouse=True)
 async def prepare_database():
-    # if settings.MODE != "TEST":
-    #     raise Exception
+    if settings.MODE != "TEST":
+        raise Exception
 
     async with engine.begin() as conn:
         await conn.run_sync(Base.metadata.drop_all)
